refactor: Share code between flows for systems boundaries (#303)
* refactor: Share code between flows for systems boundaries

I found it easy to think about this as the boundaries, and where we
traverse them. This isn't the final step in sharing, but is a good
start imo.

TOWARDS PLA-481

* Refactor/further deduplications (#311)

* Attempting further deduplications.

* Reverting change to the partial updates.

* Moving functions for PR clarity.

* Further refactors.

* Deduplicating further by parameterising on partial update callable.

* Fixing failing tests.

---------


---------
